pyproject.toml
[project]
name = "accountant"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "SEBROVATER", email = "sebrovskiy.k@gmail.com"},
]
dependencies = [
    "fastapi[full]>=0.109.2",
    "jinja2>=3.1.3",
    "tortoise-orm[aiosqlite]>=0.20.0",
    "aiohttp>=3.9.3",
    "uvicorn[standard]>=0.27.1",
    "httpx>=0.26.0",
    "pydantic-settings>=2.2.0",
    "jinja2-fragments>=1.3.0",
    "aerich>=0.7.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "GPL-3.0"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.2.1",
    "pre-commit>=3.6.0",
    "pytest>=8.0.0",
]

[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
